Configuration for using Jekyll Documentation Theme.
Added configuration to use Jekyll Docuemntation Theme,
in a way that should run directly on GitHub Pages,
so that we now have the tutorial available as
https://chaoss.github.io/grimoirelab-tutorial.

I've tried not to break the GitBook build, but in any
case this is the first commit to start a path for
moving to GitHub Pages as the primary reference for
the GrimoireLab Tutorial.

Some minor tweeks have been done to the theme, to
(hopefully) better suit the needs of the tutorial.

Some minor changes to the tutorial itself have been added
too.

To check how the tutorial looks like in GitHub Pages,
check out https://jgbarah-chaoss.github.io/grimoirelab-tutorial