update the schema separation to use double slash
https:/ I feel sure is not what was intended, but I don't blame anybody for not having automated a checker within this string or anything. Please feel free to reject if I've missed some specialization within the project or messed up the contribution. Felt silly to try to document as an issue vs. just raising a PR.